Fixed interface methods, moved jsonconverter to separate file
