+++ src/Diagnostics.ModelsAndUtils/Models/ChangeAnalysis/ResourceChangesResponseModel.cs
@@ -5,7 +5,7 @@ using Newtonsoft.Json.Linq;
 namespace Diagnostics.ModelsAndUtils.Models.ChangeAnalysis
 {
     /// <summary>
-    /// This model represents the changes data captured for a changeset in ARM Resource
+    /// This model represents the changes data captured for a changeset in ARM Resource.
     /// </summary>
     public class ResourceChangesResponseModel
     {
@@ -38,13 +38,13 @@ namespace Diagnostics.ModelsAndUtils.Models.ChangeAnalysis
         /// Old value.
         /// </summary>
         [JsonConverter(typeof(SingleOrArrayConverter<object>))]
-        public string OldValue;
+        public object OldValue;
 
         /// <summary>
         /// New value.
         /// </summary>
         [JsonConverter(typeof(SingleOrArrayConverter<object>))]
-        public string NewValue;
+        public object NewValue;
 
         /// <summary>
         /// Initiated By. It can be email address or Guid.
@@ -72,35 +72,5 @@ namespace Diagnostics.ModelsAndUtils.Models.ChangeAnalysis
         Noise = 0,
         Normal,
         Important
-    }
-
-    /// <summary>
-    /// Custom JsonConverter used to convert single or array for the same json property.
-    /// </summary>
-    internal class SingleOrArrayConverter<T> : JsonConverter
-    {
-        public override object ReadJson(JsonReader reader, Type objectType, object existingValue, JsonSerializer serializer)
-        {
-            JToken token = JToken.Load(reader);
-            if (token.Type == JTokenType.Array)
-            {
-                return token.ToObject<List<T>>();
-            }
-            return new List<T> { token.ToObject<T>() };
-        }
-
-        public override bool CanWrite
-        {
-            get { return false; }
-        }
-        public override void WriteJson(JsonWriter writer, object value, JsonSerializer serializer)
-        {
-            throw new NotImplementedException();
-        }
-
-        public override bool CanConvert(Type objectType)
-        {
-            return (objectType == typeof(List<T>));
-        }
     }
 }
