+++ src/Diagnostics.ModelsAndUtils/Models/ChangeAnalysis/SingleOrArrayConverter{T}.cs
@@ -1,0 +1,39 @@
+﻿using System;
+using System.Collections.Generic;
+using Newtonsoft.Json;
+using Newtonsoft.Json.Linq;
+
+namespace Diagnostics.ModelsAndUtils.Models.ChangeAnalysis
+{
+
+    /// <summary>
+    /// Custom JsonConverter used to convert single or array for the same json property.
+    /// </summary>
+    public class SingleOrArrayConverter<T> : JsonConverter
+    {
+        public override object ReadJson(JsonReader reader, Type objectType, object existingValue, JsonSerializer serializer)
+        {
+            JToken token = JToken.Load(reader);
+            if (token.Type == JTokenType.Array)
+            {
+                return token.ToObject<List<T>>();
+            }
+            return new List<T> { token.ToObject<T>() };
+        }
+
+        public override bool CanWrite
+        {
+            get { return false; }
+        }
+
+        public override void WriteJson(JsonWriter writer, object value, JsonSerializer serializer)
+        {
+            throw new NotImplementedException();
+        }
+
+        public override bool CanConvert(Type objectType)
+        {
+            return (objectType == typeof(List<T>));
+        }
+    }
+}
